Add proper support for multiline autofills
Autofill popups should be single-line; otherwise their size may be arbitrarily
big, which is not desired. It also improves the feel of consistency and allows
for the current simple "close" icons in the popups.

This commit adds logic that truncates any multiline replacement into its first
non-whitespace-only line before inserting it into the popup. If the autofill
replacement has any extra line(s) containing non-whitespace characters, a
truncation indicator (currently "…") is added so the user can tell that the
actual replacement will contain more than what is shown in the popup.