Improve autofill backend
This commits makes the autofill backend more general and versatile. To create a
shortcut autofill, which will display a popup and will not be applied until the
user types a trigger suffix (e.g. a space) after the autofill trigger, one must
now use ShortcutAutoFill instead of SimpleAutoFill. The thought behind this is
that a regular SimpleAutoFill, for example an automatic `)` after a `(`, is
distinctly different from a shortcut autofill, for example
`System.out.println();` instead of `syso`. The restructuring has the additional
benefit that it is no longer necessary to hardcode a space after the actual
trigger on instantiation of a shortcut autofill (`"syso "`).

The autofill backend has been modified to understand this new structure.

Modification of the autofill trigger suffix can easily be done in
ShortcutAutoFill and is has no length restriction.