Major restructuring of settings
As of this commit, we have the SettingsActivity and then we have the abstract
class SettingsFragment. Each category of settings should be implemented with a
subclass of SettingsFragment, e.g. SettingsFragmentGeneral. That subclass
should then take care of its own business, such as updating its GUI when its
settings changes. I don't yet know how this should be implemented.