Make action button bindings + labels parameterized
The bindings and labels for the action buttons were heavily hardcoded and could
not be easily modified. As of this commit, it is extremely trivial to configure
what action buttons should be attached to the LEFT and RIGHT mother buttons,
and their order.

I have also done my best to ensure that for example adding new action buttons
will be safe and easy. This is reflected in my action button handler, which
does not assume that there can be only LEFT or RIGHT alignment. If, for
example, one would add a CENTER action button, the handler will log a useful
message about its inability to recognize CENTER alignment, and then it will do
nothing, which is expected.

Also, if a button is assigned an action, for example CUT, for which nothing is
specified to happen, a helpful log entry will be issued.