Improve copyToClipboard() (not copy empty string)
The major improvement is that we will not copy anything if no text is selected,
since it makes absolutely no sense to replace potentially valuable clipboard
content with the empty string.

I have also moved the hardcoded clipboard label to strings.xml for
localizability and created a DEFAULT_LABEL for maintainability.