Simplify action button creation
Before this commit, RadialActionMenuLayout had a field holding the number of
buttons it had, which meant it was restricted to exactly three buttons, no
more, no less. It also meant that we had to perform extensive logic checking
that the number of labels being assigned to the buttons would match the number
of buttons and so on.

As of this commit, we never have to worry about the number of buttons. We can
list as many or as few actions as we want in TextEditorFragment and an action
button will be created for each one of them.

Please note that setButtons() must be called for any buttons to be created.