Add dpToCharacters() in SensitivityUtil
This function can be used to calculate how many characters the caret should
move when the user moves their finger a certain distance on the screen. In
addition to said distance (in dp), the value of the sensitivity preference
slider (a double between 0 and 1, inclusively) must be passed to the function.