Change action menu detection to use angles
The user often missed the action button when the functionality only
checked if the finger was above the button. Now it instead checks if the
fingerpress is in the correct angle. Note that this functionality might
need some tuning to be correct, now it seems to have a small
missdirection (if you try hard you notice). This works better than the
previous version though.