Fix autofillpopup crash bug
When the screen rotated the AutoFillPopupWindow was recreated to fit the
Android life cycle convention. The textfield containing the code was also
recreated and filled with the content from the current file. In this
process, there were a possibility that the textfield had no graphical
appearance but still was fed with new inputs (from the model). Therefore,
the AutoFillPopupWindow must take this in concern and check if the layout
has been created since no action should be taken if this is not the case
(it's all about the UI).