Remake of action menus
No listener is needed for button clicks anymore, instead all buttons are
connected to an instance of the interface Action. Through this interface
you can specify the name/icon of the button and what should be done when
the button fires an event. This means that the reusage of actions will be
easy.